Make GotoNextModelAdmin re-usable base class
- Use admin.ModelAdmin for better typing support
