+++ per/admin.py
@@ -1,9 +1,8 @@
 from django.contrib import admin
 from lang.admin import TranslationAdmin, TranslationInlineModelAdmin
 import per.models as models
-from per.admin_classes import RegionRestrictedAdmin, GotoNextAdminMixin
+from per.admin_classes import RegionRestrictedAdmin, GotoNextModelAdmin
 from reversion_compare.admin import CompareVersionAdmin
-from django.shortcuts import redirect
 
 
 class FormDataInline(admin.TabularInline, TranslationInlineModelAdmin):
@@ -156,31 +155,7 @@ class FormComponentQuestionAndAnswerAdmin(admin.ModelAdmin):
     pass
 
 
-class OpsLearningAdmin(admin.ModelAdmin, GotoNextAdminMixin):
-
-    def response_change(self, request, obj):
-        """Determines the HttpResponse for the change_view stage."""
-        if '_save_next' in request.POST:
-            next_pk = self.get_next_instance_pk(request, obj)
-            if next_pk:
-                response = redirect('admin:per_opslearning_change', next_pk)
-                qs = request.GET.urlencode()  # keeps _changelist_filters
-            else:
-                # Last item (or no longer in list) - go back to list in the same position
-                response = redirect('admin:per_opslearning_changelist')
-                qs = request.GET.get('_changelist_filters')
-
-            if qs:
-                response['Location'] += '?' + qs
-            return response
-
-        return super().response_change(request, obj)
-# stackoverflow.com/questions/58014139/how-to-go-to-next-object-in-django-admin
-
-    def render_change_form(self, request, context, add=False, change=False, form_url='', obj=None):
-        context.update({'show_save_and_next': True})
-        return super().render_change_form(request, context, add, change, form_url, obj)
-# stackoverflow.com/questions/49560378/cannot-hide-save-and-add-another-button-in-django-admin
+class OpsLearningAdmin(GotoNextModelAdmin):
 
     search_fields = ("learning", "learning_validated")
     list_filter = ("sector", "sector_validated", "per_component", "per_component_validated")
